app/indexer.py
import faiss
import numpy as np

def build_faiss_index(embeddings: list[list[float]]) -> faiss.IndexFlatL2:
    dim = len(embeddings[0])
    index = faiss.IndexFlatL2(dim)
    index.add(np.array(embeddings).astype('float32'))
    return index